requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.7
django-appconf==1.0.5
django-celery-beat==2.3.0
django-compressor==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-postgres-extra==2.0.6
django-sass-processor==1.2.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-friendly-errors-django-4==0.15.2
flower==1.2.0
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
hiredis==2.0.0
humanize==4.3.0
idna==3.4
kombu==5.2.4
libsass==0.21.0
packaging==21.3
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
PyJWT==2.5.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
rcssmin==1.1.0
redis==4.3.4
requests==2.28.1
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.2
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xml-to-dict==0.1.6
zope.event==4.5.0
zope.interface==5.4.0
